fix: move getUserPisteCredentials to lib (Next.js route files cannot have non-HTTP exports)
