+++ src/app/api/settings/test-piste/route.ts
@@ -2,7 +2,7 @@ import { NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
 import { authOptions } from '@/lib/auth-options';
 import { testPisteConnection } from '@/lib/piste-api';
-import { getUserPisteCredentials } from '@/app/api/settings/piste-credentials/route';
+import { getUserPisteCredentials } from '@/lib/user-piste-credentials';
 import { decryptCredential } from '@/lib/credential-crypto';
 
 // POST — test PISTE connection using body credentials (in-flight) or saved DB credentials
